- add employee_test.py

+++ northwind/employee/models.py
@@ -28,7 +28,7 @@ class Territories(models.Model):
 
 
 class Employees(models.Model):
-    employee_id = models.SmallIntegerField(primary_key=True)
+    employee_id = models.AutoField(primary_key=True)
     last_name = models.CharField(max_length=20)
     first_name = models.CharField(max_length=10)
     title = models.CharField(max_length=30, blank=True, null=True)
